requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
requests==2.32.3
python-dotenv==1.0.1
PyGithub==2.3.0
mcp==1.0.0
httpx==0.27.2
pydantic==2.8.2
